requirements.txt
Appium-Python-Client==4.5.0
pytest==7.2.0
python-dotenv==1.0.0